@@ -2,21 +2,23 @@ source 'http://rubygems.org'
 
 ruby '2.4.5'
 
-gem 'rails', '4.2.11'
+gem 'rails', '5.2.2.1'
 gem 'rake', '< 11.0'
+gem 'bootsnap','>= 1.1.0', require: false
 
-gem 'country_select', '~> 1.1.3'
 gem 'pg', '~> 0.20.0'
-gem 'textacular', '~> 4.0', :require => 'textacular/searchable'
-gem 'jquery-rails', '~> 2.1.0'
-gem 'clearance', '~> 1.5.0'
+gem 'textacular', '~> 5.1.0'
+gem 'jquery-rails', '~> 4.1.1'
+gem 'jquery-ui-rails', '~> 3.0.1'
+gem 'clearance', '~> 1.13.0'
+gem 'clearance-deprecated_password_strategies'
 gem 'high_voltage', '~> 1'
 gem 'paperclip', '~> 4.1.1'
 gem 'fog', '~> 1.41.0'
 gem 'formtastic'
 gem 'flutie'
 gem 'bourbon', '~> 4.0.2'
-gem "simple_form", "~> 3.0"
+gem 'simple_form', '~> 4.1'
 gem 'nokogiri', '~> 1.10.1'
 gem "will_paginate", "~> 3.1.6"
 gem "friendly_id", "~> 5.2.4"
@@ -36,8 +38,9 @@ gem 'xmlrpc'
 
 group :development do
   gem "letter_opener"
-  gem 'thin'
-  gem 'web-console', '~> 2.0'
+  gem "listen"
+  gem 'puma'
+  gem 'web-console'
 end
 
 group :development, :test do
@@ -51,14 +54,16 @@ end
 
 group :test do
   gem "turnip"
-  gem "capybara"
+  gem "capybara", "~> 2.18.0"
+  gem "capybara-screenshot", "~> 1.0.11"
+  gem "chromedriver-helper"
   gem "database_cleaner"
-  gem "capybara-webkit", "~> 1.3.0"
-  gem 'capybara-screenshot'
   gem "factory_girl_rails"
   gem "faker"
   gem "bourne"
   gem "timecop"
+  gem "rails-controller-testing"
+  gem "selenium-webdriver"
   gem "shoulda-matchers"
   gem "launchy"
   gem "email_spec"
