@@ -1,4 +1,4 @@
-class RenameSubmissionsToProjects < ActiveRecord::Migration
+class RenameSubmissionsToProjects < ActiveRecord::Migration[4.2]
   def change
     rename_table :projects, :accepted_projects
     rename_table :submissions, :projects
