@@ -1,4 +1,4 @@
-class CreateVotes < ActiveRecord::Migration
+class CreateVotes < ActiveRecord::Migration[4.2]
   def change
     create_table :votes do |t|
       t.references :user
