@@ -1,4 +1,4 @@
-class LastChapterUserIsViewing < ActiveRecord::Migration
+class LastChapterUserIsViewing < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :last_viewed_chapter_id, :int
   end
