@@ -1,4 +1,4 @@
-class CreatePhotos < ActiveRecord::Migration
+class CreatePhotos < ActiveRecord::Migration[4.2]
   def change
     create_table :photos do |t|
       t.timestamps
