@@ -1,4 +1,4 @@
-class AddEmailAddressToChapter < ActiveRecord::Migration
+class AddEmailAddressToChapter < ActiveRecord::Migration[4.2]
   def change
     add_column :chapters, :email_address, :string
   end
