@@ -1,4 +1,4 @@
-class AddFundedDescription < ActiveRecord::Migration
+class AddFundedDescription < ActiveRecord::Migration[4.2]
   def up
     add_column "projects", "funded_description", :text
     Project.reset_column_information
