@@ -1,4 +1,4 @@
-class AddChapterTimeZone < ActiveRecord::Migration
+class AddChapterTimeZone < ActiveRecord::Migration[4.2]
   def change
     add_column 'chapters', 'time_zone', :string, :limit => 50
   end
