@@ -1,15 +1,11 @@
-$('a.chapters').toggle(function() {
-  $('.chapter-menu').addClass("expanded");
-  $('a.chapters').addClass("active");
-}, function() {
-  $('.chapter-menu').removeClass("expanded");
-  $('a.chapters').removeClass("active");
+$('a.chapters').on('click', function(event) {
+  event.preventDefault();
+  $('.chapter-menu').toggleClass('expanded');
+  $('a.chapters').toggleClass('active');
 });
 
-$('a.chapter-selection').toggle(function() {
-  $('.chapter-selector').addClass("expanded");
-  $('a.chapter-selection').addClass("expanded");
-}, function() {
-  $('.chapter-selector').removeClass("expanded");
-  $('a.chapter-selection').removeClass("expanded");
+$('a.chapter-selection').on('click', function(event) {
+  event.preventDefault();
+  $('.chapter-selector').toggleClass('expanded');
+  $('a.chapter-selection').toggleClass('expanded');
 });
