@@ -15,7 +15,9 @@ end
 step 'I set the last image to be first' do
   photo_ids = @project.photos.map(&:id)
   # the input is hidden, but it's easier to set it than to deal with drag/drop
-  find("#project_photo_order", visible: false).set([photo_ids[2], photo_ids[0], photo_ids[1]].join(" "))
+  # find("#project_photo_order", visible: false).set([photo_ids[2], photo_ids[0], photo_ids[1]].join(" "))
+  # Selenium doesn't let us manipulate a hidden element, so we set it with javascript instead
+  page.execute_script("document.getElementById('project_photo_order').value = '#{[photo_ids[2], photo_ids[0], photo_ids[1]].join(' ')}'");
   click_button("Save")
 end
 
