@@ -2,6 +2,7 @@
 ENV["RAILS_ENV"] ||= 'test'
 require File.expand_path("../../config/environment", __FILE__)
 require 'rspec/rails'
+require 'clearance/rspec'
 require 'turnip'
 require 'turnip/capybara'
 require 'database_cleaner'
@@ -9,8 +10,6 @@ require 'paperclip/matchers'
 require 'sucker_punch/testing/inline'
 
 DatabaseCleaner.strategy = :truncation
-Capybara.javascript_driver = :webkit
-require 'capybara-screenshot/rspec'
 
 # Requires supporting ruby files with custom matchers and macros, etc,
 # in spec/support/ and its subdirectories.
@@ -18,8 +17,6 @@ Dir[Rails.root.join("spec/support/**/*.rb")].each {|f| require f}
 Dir[Rails.root.join("spec/features/step_definitions**/*.rb")].each {|f| require f}
 
 RSpec.configure do |config|
-  config.include Capybara::DSL
-
   config.infer_spec_type_from_file_location!
 
   config.mock_with :mocha
